pyproject.toml
[project]
name = "funcproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-functions>=1.24.0",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
markers = [
    "integration: Integration tests requiring Azure Functions runtime (run with: pytest -m integration)",
]
testpaths = ["tests"]
